Refactor NonTerminatorInsn class
