+++ compiler/codegen/BalNonTerminatorInsn.cpp
@@ -1,8 +1,10 @@
-#include "BIR.h"
+#include "BalNonTerminatorInsn.h"
 
+namespace nballerina {
 NonTerminatorInsn::NonTerminatorInsn(Location *posi, InstructionKind kindP,
                                      Operand *lOp)
     : AbstractInsn(posi, kindP, lOp) {}
 
 void NonTerminatorInsn::translate(__attribute__((unused))
                                   LLVMModuleRef &modRef) {}
+} // namespace nballerina
